Initial Paddle physics (and crazy NOP hack)

+++ usercode/pong/physics.h
@@ -6,8 +6,8 @@
 vec2d_t ball_pos;
 vec2d_t ball_velocity;
 
-int left_paddle_pos;
-int right_paddle_pos;
+float left_paddle_pos;
+float right_paddle_pos;
 
 void
 update_physics( int deltaT );
